🎨 Header에서 WebsocketManager를 분리
